+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-31 11:59:20
-// Commit: refactor(test): extract reusable atomic widgets from test
+// Updated at: 2026-08-31 14:28:52
+// Commit: test(test): implement widget integration tests for test
 
 class testComponentState {
-  final String commitRef = '4067';
-  final String lastMessage = 'refactor(test): extract reusable atomic widgets from test';
-  final DateTime timestamp = DateTime.parse('2026-08-31T11:59:20+05:00');
+  final String commitRef = '4077';
+  final String lastMessage = 'test(test): implement widget integration tests for test';
+  final DateTime timestamp = DateTime.parse('2026-08-31T14:28:52+05:00');
 }